% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_sets.R
\name{create_conditioning_sets}
\alias{create_conditioning_sets}
\title{Conditioning Sets Function}
\usage{
create_conditioning_sets(p)
}
\arguments{
\item{p}{is the number of nodes in the graph}
}
\description{
This function creates a list of lists for each node, with each node's list containing
lists that contain vectors indicating the separation sets for the two nodes.
}
