% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbors.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Find Neighbors}
\usage{
find_neighbors(t, true_dag)
}
\arguments{
\item{t}{the node of interest}

\item{true_dag}{provides the adjacency matrix for the true DAG
Note for the DAG adjacency matrix: A_{ij}=1 if i -> j}
}
\description{
This function finds all the neighbors for a particular node.
A neighbor is defined as the set of parents, children, and spouses of the node.
}
