% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pval.R
\name{get_pval}
\alias{get_pval}
\title{Get p-value for Node Correlation}
\usage{
get_pval(var_list)
}
\arguments{
\item{cor.mat}{the correlation matrix for all the nodes}

\item{i}{the potential parent node}

\item{j}{the potential child node}

\item{k}{the potential separating set}
}
\description{
Get p-value for Node Correlation
}
