% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_pc2.R
\name{local_pc2}
\alias{local_pc2}
\title{Local PC Algorithm (Efficient)}
\usage{
local_pc2(data = NULL, true_dag = NULL, target, pop = TRUE,
  verbose = TRUE, verbose_small = TRUE)
}
\arguments{
\item{data}{default is NULL, the sample version when we have data}

\item{true_dag}{given for the population version}

\item{target}{vector of nodes of interest}

\item{cor.mat}{default is NULL, the correlation matrix for the nodes (population version)}
}
\description{
Local PC Algorithm (Efficient)
}
