\name{LocalPC-package}
\alias{LocalPC-package}
\alias{LocalPC}
\docType{package}
\title{\packageTitle{LocalPC}}
\description{\packageDescription{LocalPC}}
\section{Package Content}{\packageIndices{LocalPC}}
\author{\packageAuthor{LocalPC}}
\section{Maintainer}{\packageMaintainer{LocalPC}}
\keyword{package}
