% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_skel_loc.R
\name{pc_skel_loc}
\alias{pc_skel_loc}
\title{PC Algorithm for Skeleton (for Local PC)}
\usage{
pc_skel_loc(dataset = NULL, C_tilde = NULL, true_dag = NULL,
  neighbors = NULL, pop = TRUE, lmax = 3, verbose = TRUE,
  tol = 0.005)
}
\arguments{
\item{dataset}{default is NULL, only used when we are using sample version and have data}

\item{C_tilde}{default is NULL, used in the local algorithm with each clique being considered forming a complete graph}

\item{lmax}{the maximum size of the separating sets that we will consider (default is 3)}

\item{verbose}{messages concerning the operation of the algorithm (default is TRUE)}

\item{tol}{the significance level (default is 0.005)}

\item{cor.mat}{default is NULL, the correlation matrix for the nodes (population algorithm)}
}
\description{
PC Algorithm for Skeleton (for Local PC)
}
