% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_vstruct.R
\name{pc_vstruct}
\alias{pc_vstruct}
\title{Obtain v-structures}
\usage{
pc_vstruct(G, S, verbose = TRUE)
}
\arguments{
\item{G}{the current graph after constructing the skeleton}

\item{S}{the list of separating sets for each node}
}
\description{
After obtaining the skeleton from the first step of the PC algorithm
we use one of Meek's rules to orient the v-structures
}
